requirements.txt
django
django-crispy-forms
crispy_bootstrap5
openai
python-decouple
nibabel
numpy
psycopg2
boto3
scipy
numba
nilearn
tqdm
natsort
importlib
pfc-toolkit
django-storages
pillow
